PR comments update
