+++ aws/rust-runtime/aws-config/Cargo.toml
@@ -35,6 +35,8 @@ aws-smithy-runtime = { path = "../../sdk/build/aws-sdk/sdk/aws-smithy-runtime", 
 aws-smithy-runtime-api = { path = "../../sdk/build/aws-sdk/sdk/aws-smithy-runtime-api", features = ["client"] }
 aws-smithy-types = { path = "../../sdk/build/aws-sdk/sdk/aws-smithy-types" }
 aws-types = { path = "../../sdk/build/aws-sdk/sdk/aws-types" }
+bytes = "1.1.0"
+http = "0.2.4"
 hyper = { version = "0.14.26", default-features = false }
 time = { version = "0.3.4", features = ["parsing"] }
 tokio = { version = "1.13.1", features = ["sync"] }
@@ -43,9 +45,6 @@ url = "2.3.1"
 
 # implementation detail of IMDS credentials provider
 fastrand = "2.0.0"
-
-bytes = "1.1.0"
-http = "0.2.4"
 
 # implementation detail of SSO credential caching
 aws-sdk-sso = { path = "../../sdk/build/aws-sdk/sdk/sso", default-features = false, optional = true }
@@ -62,11 +61,7 @@ aws-smithy-runtime-api = { path = "../../sdk/build/aws-sdk/sdk/aws-smithy-runtim
 futures-util = { version = "0.3.29", default-features = false }
 tracing-test = "0.2.4"
 tracing-subscriber = { version = "0.3.16", features = ["fmt", "json"] }
-
 tokio = { version = "1.23.1", features = ["full", "test-util"] }
-
-# used for fuzzing profile parsing
-arbitrary = "1.3"
 
 # used for test case deserialization
 serde = { version = "1", features = ["derive"] }
