+++ aws/rust-runtime/aws-config/src/profile/profile_file.rs
@@ -7,18 +7,18 @@
 
 /// Use aws_runtime::env_config::file::EnvConfigFiles instead.
 #[deprecated(
-    since = "1.1.10",
+    since = "1.1.11",
     note = "Use aws_runtime::env_config::file::EnvConfigFiles instead."
 )]
 pub type ProfileFiles = aws_runtime::env_config::file::EnvConfigFiles;
 
 /// Use aws_runtime::env_config::file::Builder instead.
-#[deprecated(since = "1.1.10", note = "Use aws_runtime::env_config::file::Builder.")]
+#[deprecated(since = "1.1.11", note = "Use aws_runtime::env_config::file::Builder.")]
 pub type Builder = aws_runtime::env_config::file::Builder;
 
 /// Use aws_runtime::env_config::file::EnvConfigFileKind instead.
 #[deprecated(
-    since = "1.1.10",
+    since = "1.1.11",
     note = "Use aws_runtime::env_config::file::EnvConfigFileKind."
 )]
 pub type ProfileFileKind = aws_runtime::env_config::file::EnvConfigFileKind;
